requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
ortools==9.8.3296
supabase==2.3.0
python-dotenv==1.0.0
pytz==2024.1
pydantic==2.5.3